msm8998: Final change to move the dsp partition into A/B mode
We now move the mounting of the dsp partition into the fstab so that
fs_mgr can use the slotselect argument to correctly mount the partition
based on the current active slot.

Also moved the persist partition into fs_mgr since there is no real reason
anymore to keep it in the init.target.rc file.

Change-Id: I5a76ec3407545b14e291a78dedd0ae66cbaa1c8e