oneplus5: Reduce max_freq_hysteresis for big cluster
We need to balance max_freq_hysteresis with the
increased hispeed_freq on big cores in order to
lower the time spent at hispeed_freq and to lower
the timeframe where we can get a direct re-boost to
max_freq.

So, we reduce max_freq_hysteresis by 20ms for the
big cluster.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>