'Enable' qcomsysd service instead of 'start'
On force-encrypted devices, main class services will be
started-stopped-started again due to the process of encryption. So, on
such devices, our qcomsysd doesn't start the second time along with main
class services startup on 1st bootup, since its disabled by default and we
use 'start' to start it manually on property detection. Move to 'enable',
so that we just enable the service and init will properly start it as part
of main class start-up, so that even on force-encrypted devices,
it will run properly on 1st boot.
Also, stop it explicitly as well, if property value is 0.

Change-Id: I12d11be9b4ab017f177140c4538edc2664e57427