oneplus5(t): Add config for display composition color spaces
Currently we use the Display-P3 color space for display composition
across all color modes on OP6. We do this to facilitate  a new version
of Adaptive color mode; this forces display to always use P3 color
calibration.

It turns out that in OP6's case using P3 color calibration for sRGB
content, even when the content is accurately converted to P3, degrades
color accuracy for certain sRGB colors. To fix this, we revert back to
pre-OP6 behavior, where we let SurfaceFlinger choose the most
appropriate color space for composition, for all color modes used on OP6
*except* for Adaptive color mode.

Bug: 137140317
Bug: 137053654
Change-Id: Id9c414de862000c59d860776277df1cb0a719b06