Update patch set 2
Patch Set 2:

Not sure on post-fs-data in init.pa.rc is late enough, you should verify this with adb shell. If it's not, you can either add a new trigger in the rc or just move it to post boot and convert to shell.

Patch-set: 2
Reviewer: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>