Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I don't think walleye did this and I don't think you have this node yet on your device. I didn't see a commit to enable the defconfig for it.

we have this node already that i've verified and it's set by default with the values you've grabbed from crosshatch.
this current values were fetched from latest 10 for Pixel 2 where OnePlus has the same values in system/init.rc . Although in the newest Android R for Pixel 2 the values are back from the ones you have in crosshatch..

Still don't know though why although is being set the stock kernel doesn't seem to have the defconfig for it. Will need to make proper tests on this one for sure

Patch-set: 1